Updated to 2021

+++ src/main/java/frc/robot/subsystems/HangerArm.java
@@ -28,7 +28,7 @@ public final class HangerArm extends SubsystemBaseWrapper implements MotorSubsys
         motor.set(ControlMode.PercentOutput, 0);
     }
     public int getEncoderVale(){
-        return motor.getSelectedSensorPosition();
+        return (int) motor.getSelectedSensorPosition();
     }
     public void setEncoderValue(int value){
         motor.setSelectedSensorPosition(value);
